+++ src/main/java/com/commercetools/sync/inventories/utils/InventoryReferenceReplacementUtils.java
@@ -1,0 +1,38 @@
+package com.commercetools.sync.inventories.utils;
+
+import io.sphere.sdk.inventory.InventoryEntry;
+import io.sphere.sdk.inventory.InventoryEntryDraft;
+import io.sphere.sdk.inventory.InventoryEntryDraftBuilder;
+import io.sphere.sdk.types.CustomFieldsDraft;
+
+import javax.annotation.Nonnull;
+import java.util.List;
+import java.util.stream.Collectors;
+
+import static com.commercetools.sync.commons.utils.SyncUtils.replaceCustomTypeIdWithKeys;
+
+public final class InventoryReferenceReplacementUtils{
+    /**
+     * Takes a list of inventoryEntries that are supposed to have their custom type reference expanded
+     * in order to be able to fetch the keys and replace the reference ids with the corresponding keys and then return
+     * a new list of inventory entry drafts with their references containing keys instead of the ids.  Note that if the
+     * references are not expanded for an inventory entry, the reference ids will not be replaced with keys and will
+     * still have their ids in place.
+     *
+     * @param inventoryEntries the inventory entries to replace their reference ids with keys
+     * @return a list of inventoryEntry drafts with keys instead of ids for references.
+     */
+    @Nonnull
+    public static List<InventoryEntryDraft> replaceInventoriesReferenceIdsWithKeys(@Nonnull final List<InventoryEntry>
+                                                                                       inventoryEntries) {
+        return inventoryEntries
+            .stream()
+            .map(inventoryEntry -> {
+                final CustomFieldsDraft customTypeWithKeysInReference = replaceCustomTypeIdWithKeys(inventoryEntry);
+                return InventoryEntryDraftBuilder.of(inventoryEntry)
+                                                 .custom(customTypeWithKeysInReference)
+                                                 .build();
+            })
+            .collect(Collectors.toList());
+    }
+}